pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dk1"
version = "2025.8.0"
dependencies = [
  "numpy",
  "pyserial",
]

authors = [
  {name = "Jannik Grothusen", email = "jannik.grothusen@robot-learning.co"},
]
maintainers = [
  {name = "Jannik Grothusen", email = "jannik.grothusen@robot-learning.co"},
]

[project.urls]
Homepage = "https://example.com"
Documentation = "https://readthedocs.org"
Repository = "https://github.com/me/spam.git"
